删除exe_task
