+++ src-tauri/src/mxu_actions.rs
@@ -834,141 +834,6 @@ fn execute_power_sleep() -> bool {
 }
 
 // ============================================================================
-// MXU_EXEC_TASK Custom Action
-// ============================================================================
-
-/// MXU_EXEC_TASK 动作名称常量
-const MXU_EXEC_TASK_ACTION: &str = "MXU_EXEC_TASK_ACTION";
-
-/// MXU_EXEC_TASK custom action 回调函数
-/// 从 custom_action_param 中读取 exec、args、cwd，启动外部程序并等待退出。
-///
-/// 子进程 stdout 会被管道化并逐行转发到 `maa-agent-output` 通道（复用 Agent 输出
-/// 管线），从而显示在 MXU 运行日志中。Windows 下隐藏子进程控制台窗口。
-fn mxu_exec_task_action_impl(
-    _ctx: &maa_framework::context::Context,
-    args: &maa_framework::custom::ActionArgs,
-    app_handle: &AppHandle,
-    instance_id: &str,
-) -> bool {
-    let param_str = args.param;
-    info!("[MXU_EXEC_TASK] Received param: {}", param_str);
-
-    let json: serde_json::Value = match serde_json::from_str(param_str) {
-        Ok(v) => v,
-        Err(e) => {
-            warn!("[MXU_EXEC_TASK] Failed to parse param JSON: {}", e);
-            return false;
-        }
-    };
-
-    let exec = match json.get("exec").and_then(|v| v.as_str()) {
-        Some(p) if !p.trim().is_empty() => p.to_string(),
-        _ => {
-            warn!("[MXU_EXEC_TASK] Missing or empty 'exec' parameter");
-            return false;
-        }
-    };
-
-    let args_vec: Vec<String> = json
-        .get("args")
-        .and_then(|v| v.as_array())
-        .map(|arr| {
-            arr.iter()
-                .filter_map(|v| v.as_str().map(|s| s.to_string()))
-                .collect()
-        })
-        .unwrap_or_default();
-
-    let cwd = json
-        .get("cwd")
-        .and_then(|v| v.as_str())
-        .filter(|s| !s.trim().is_empty())
-        .map(|s| s.to_string());
-
-    info!(
-        "[MXU_EXEC_TASK] Running: exec={}, args={:?}, cwd={:?}",
-        exec, args_vec, cwd
-    );
-
-    let mut cmd = crate::commands::utils::build_launch_command(&exec, &args_vec, false);
-
-    if let Some(ref dir) = cwd {
-        if std::path::Path::new(dir).exists() {
-            cmd.current_dir(dir);
-        } else {
-            warn!("[MXU_EXEC_TASK] CWD does not exist: {}", dir);
-        }
-    } else if let Some(parent) = std::path::Path::new(&exec).parent() {
-        if parent.exists() {
-            cmd.current_dir(parent);
-        }
-    }
-
-    // 覆盖 stdout 为管道，用于把 go-service 的用户日志转发到运行日志。
-    // stderr 保持 null：go-service 已将自身 stderr 重定向到文件。
-    cmd.stdout(std::process::Stdio::piped());
-
-    // Windows 下隐藏子进程控制台窗口（go-service 是控制台子系统程序）。
-    #[cfg(target_os = "windows")]
-    {
-        use std::os::windows::process::CommandExt;
-        const CREATE_NO_WINDOW: u32 = 0x0800_0000;
-        cmd.creation_flags(CREATE_NO_WINDOW);
-    }
-
-    let mut child = match cmd.spawn() {
-        Ok(c) => c,
-        Err(e) => {
-            log::error!("[MXU_EXEC_TASK] Failed to spawn program: {}", e);
-            return false;
-        }
-    };
-
-    // 单独线程按行读取 stdout，转发到 maa-agent-output 通道（复用 Agent 输出管线，
-    // 自带批处理与前端刷屏抑制）。
-    if let Some(stdout) = child.stdout.take() {
-        let batcher = crate::commands::maa_agent::AgentOutputBatcher::new(
-            app_handle.clone(),
-            instance_id.to_string(),
-        );
-        std::thread::spawn(move || {
-            use std::io::{BufRead, BufReader};
-            let mut reader = BufReader::new(stdout);
-            let mut buffer = Vec::new();
-            loop {
-                buffer.clear();
-                match reader.read_until(b'\n', &mut buffer) {
-                    Ok(0) => break,
-                    Ok(_) => {
-                        let line = String::from_utf8_lossy(&buffer);
-                        batcher.enqueue("stdout", line.trim_end());
-                    }
-                    Err(_) => break,
-                }
-            }
-        });
-    }
-
-    match child.wait() {
-        Ok(status) => {
-            let exit_code = status.code().unwrap_or(-1);
-            if exit_code == 0 {
-                info!("[MXU_EXEC_TASK] Process exited successfully");
-                true
-            } else {
-                warn!("[MXU_EXEC_TASK] Process exited with code: {}", exit_code);
-                false
-            }
-        }
-        Err(e) => {
-            log::error!("[MXU_EXEC_TASK] Failed to wait for program: {}", e);
-            false
-        }
-    }
-}
-
-// ============================================================================
 // 注册入口
 // ============================================================================
 
@@ -1017,46 +882,6 @@ pub fn register_all_mxu_actions(
     reg_action!(MXU_WEBHOOK_ACTION, mxu_webhook_action_fn);
     reg_action!(MXU_NOTIFY_ACTION, mxu_notify_action_fn);
     reg_action!(MXU_POWER_ACTION, mxu_power_action_fn);
-
-    // MXU_EXEC_TASK 需要 app_handle + instance_id 才能把子进程 stdout 转发到运行日志，
-    // 因此与 killproc 一样单独注册（无法走上面的 reg_action! 宏）。
-    let exec_task_app_handle = app_handle.clone();
-    let exec_task_instance_id = instance_id.to_string();
-    let exec_task_wrapper = move |ctx: &maa_framework::context::Context,
-                                  args: &maa_framework::custom::ActionArgs|
-          -> bool {
-        std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-            mxu_exec_task_action_impl(ctx, args, &exec_task_app_handle, &exec_task_instance_id)
-        }))
-        .unwrap_or_else(|e| {
-            let msg = if let Some(s) = e.downcast_ref::<&str>() {
-                s.to_string()
-            } else if let Some(s) = e.downcast_ref::<String>() {
-                s.clone()
-            } else {
-                "Unknown panic payload".to_string()
-            };
-            log::error!(
-                "[MXU] Custom action {} panicked: {}",
-                MXU_EXEC_TASK_ACTION,
-                msg
-            );
-            false
-        })
-    };
-
-    if let Err(e) = resource.register_custom_action(
-        MXU_EXEC_TASK_ACTION,
-        Box::new(FnAction::new(exec_task_wrapper)),
-    ) {
-        warn!("[MXU] Failed to register {}: {:?}", MXU_EXEC_TASK_ACTION, e);
-        failed_count += 1;
-    } else {
-        info!(
-            "[MXU] Custom action {} registered successfully",
-            MXU_EXEC_TASK_ACTION
-        );
-    }
 
     let killproc_app_handle = app_handle.clone();
     let killproc_instance_id = instance_id.to_string();
