+++ src/components/Toolbar.tsx
@@ -16,7 +16,6 @@ import { maaService } from '@/services/maaService';
 import clsx from 'clsx';
 import { loggers, generateTaskPipelineOverride, computeResourcePaths } from '@/utils';
 import { getMxuSpecialTask } from '@/types/specialTasks';
-import { getExecTaskItem, buildExecTaskDef } from '@/types/execTasks';
 import { isPretaskName, getPretaskItem, buildPretaskArgs } from '@/types/pretasks';
 import { splitTasksIntoThreeSegments, shouldSkipScreenshot } from '@/utils/taskSegmentation';
 import type { TaskConfig, ControllerConfig } from '@/types/maa';
@@ -1079,10 +1078,8 @@ export function Toolbar({ showAddPanel, onToggleAddPanel, className }: ToolbarPr
             continue;
           }
           const specialTask = getMxuSpecialTask(selectedTask.taskName);
-          const execTaskItem = getExecTaskItem(projectInterface, selectedTask.taskName);
           const taskDef =
             specialTask?.taskDef ||
-            (execTaskItem ? buildExecTaskDef(execTaskItem) : undefined) ||
             projectInterface?.task.find((t) => t.name === selectedTask.taskName);
           if (!taskDef) {
             log.warn(`跳过任务 ${selectedTask.taskName}: 未找到任务定义`);
@@ -1129,7 +1126,6 @@ export function Toolbar({ showAddPanel, onToggleAddPanel, className }: ToolbarPr
                 projectInterface,
                 currentControllerName,
                 currentResourceName,
-                basePath,
               ),
               selected_task_id: selectedTask.id,
             };
