+++ src/i18n/locales/zh-TW.ts
@@ -482,7 +482,6 @@ export default {
     noResults: '沒有找到符合的任務',
     alreadyAdded: '已新增',
     specialTasks: '特殊任務',
-    execTasks: '輔助任務',
     pretasks: '前置任務',
     allSpecialTasksAdded: '已全部新增',
     collapse: '收起面板',
