+++ src/services/interfaceLoader.ts
@@ -6,10 +6,9 @@ import type {
   ControllerType,
   PresetItem,
   GroupItem,
-  ExecTaskItem,
   PretaskItem,
 } from '@/types/interface';
-import { normalizeExecTaskConfigs, normalizePretaskConfigs } from '@/types/interface';
+import { normalizePretaskConfigs } from '@/types/interface';
 import { loggers } from '@/utils/logger';
 import { parseJsonc } from '@/utils/jsonc';
 import { isTauri } from '@/utils/paths';
@@ -29,8 +28,6 @@ interface ImportableInterface {
   group?: GroupItem[];
   /** v2.7.0: 支持导入 pretask */
   pretask?: PretaskItem | PretaskItem[];
-  /** v2.9.0: 支持导入 exec_task */
-  exec_task?: ExecTaskItem | ExecTaskItem[];
 }
 
 export interface LoadResult {
@@ -280,14 +277,6 @@ function mergeImported(pi: ProjectInterface, imported: ImportableInterface): voi
     const existing = normalizePretaskConfigs(pi.pretask) || [];
     pi.pretask = [...existing, ...importedPretasks];
     log.info(`合并了 ${importedPretasks.length} 个导入的 pretask`);
-  }
-
-  // v2.9.0: 合并 exec_task（单对象视为一项，按导入顺序追加为有序列表）
-  const importedExecTasks = normalizeExecTaskConfigs(imported.exec_task);
-  if (importedExecTasks && importedExecTasks.length > 0) {
-    const existing = normalizeExecTaskConfigs(pi.exec_task) || [];
-    pi.exec_task = [...existing, ...importedExecTasks];
-    log.info(`合并了 ${importedExecTasks.length} 个导入的 exec_task`);
   }
 }
